requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
requests>=2.28.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.25.0
fastapi>=0.100.0
uvicorn>=0.23.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
